Made some changes on how the services are controlled fromt he rpeference page, also conversions of speeds

+++ trunk/src/com/mobilityspot/SpeedUnitsConversion.java
@@ -1,12 +1,12 @@
 package com.mobilityspot;
 
 public class SpeedUnitsConversion {
-	  public static double mphToMeterPerSecond(int mphSpeed) {
+	  public static double mphToMeterPerSecond(Double mphSpeed) {
 		double msSpeed = (mphSpeed * 1609.34) /3600; 
 		return msSpeed; 
 	  }
 	  
-	  public static double kphToMeterPerSecond(int kphSpeed) {
+	  public static double kphToMeterPerSecond(Double kphSpeed) {
 		double msSpeed = (kphSpeed * 1000) /3600; 
 		return msSpeed;  
 	  }
